% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\docType{data}
\name{filepaths}
\alias{filepaths}
\title{Get enzyme activities}
\format{
An object of class \code{character} of length 3.
}
\usage{
filepaths
}
\arguments{
\item{filepaths}{String or vector containing filepath(s) to .CSV files}

\item{dropfolder}{Optional logical for whether to keep the whole filepath or drop the folders leading up to the file.}
}
\value{
A dataframe enzyme activity in AU/s
}
\description{
Read .CSV file and calculate the slope in Absorbance Units, AU, per second. A start time must be provided.
}
\keyword{datasets}
