% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{read_activity}
\alias{read_activity}
\title{Get enzyme activities}
\usage{
read_activity(filepaths, start, end)
}
\arguments{
\item{filepaths}{String or vector containing filepath(s) to .CSV files}
}
\value{
A dataframe enzyme activity in AU/s
}
\description{
Read .CSV file and calculate the slope in Absorbance Units, AU, per second. A start time must be provided.
}
