% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_chl.R
\name{read_chl}
\alias{read_chl}
\title{Get chlorophyll concentrations}
\usage{
read_chl(filename, units = c("nmol_ml"))
}
\arguments{
\item{filename}{Filepath to a file}

\item{units}{nmol_ml or ug_ml}
}
\value{
A dataframe containing chlorophyll concentrations in specified units
}
\description{
Read .WAV file for chlorophyll concentrations in 80% aqueous acetone according to Porra et al. 1989 BBA
}
