% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{read_csv8452}
\alias{read_csv8452}
\title{Read HP8452 .CSV file}
\usage{
read_csv8452(filepath, start, end, dropfolder = TRUE, droppath = FALSE)
}
\arguments{
\item{filepath}{String containing filepath(s) to .CSV files}

\item{start}{Start of reaction. By default the function will use the lowest value of Time if missing.}

\item{end}{Start of reaction. By default the function will use the highest value of Time if missing.}

\item{dropfolder}{Optional logical for whether to keep the whole filepath or drop the folders leading up to the file.}
}
\value{
A dataframe with time points of absorbance measurements
}
\description{
Read .CSV file and add a filename column.
}
