% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{read_wav}
\alias{read_wav}
\title{Get chlorophyll concentrations}
\usage{
read_wav(filepaths, units = c("nmol_ml"))
}
\arguments{
\item{units}{nmol_ml or ug_ml or both in a vector}

\item{filepath}{String of filepath to .WAV file}
}
\value{
A dataframe containing chlorophyll concentrations in specified units
}
\description{
Read .WAV file for chlorophyll concentrations in 80% aqueous acetone according to Porra et al. 1989 BBA
}
