% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_activity.R
\name{read_activity}
\alias{read_activity}
\title{Get enzyme activity}
\usage{
read_activity(filename, start, end)
}
\arguments{
\item{filename}{Filepath to a file}
}
\value{
A dataframe enzyme activity in AU/s
}
\description{
Read .CSV file and calculate the slope in Absorbance Units, AU, per second. A start time must be provided.
}
