% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{read_chl}
\alias{read_chl}
\title{Get chlorophyll concentrations (one or more files)
Read one or more .WAV files for chlorophyll concentrations in 80% aqueous acetone according to Porra et al. 1989 BBA}
\usage{
read_chl(filepaths, units)
}
\arguments{
\item{filepaths}{String or vector of filepath(s) to .WAV file}

\item{units}{nmol_ml or ug_ml or both in a vector}
}
\value{
A dataframe containing chlorophyll concentrations in specified units
}
\description{
Get chlorophyll concentrations (one or more files)
Read one or more .WAV files for chlorophyll concentrations in 80% aqueous acetone according to Porra et al. 1989 BBA
}
